refactor(initrd): Consolidate common test variables and methods
This commit homogenizes the tests such that they all use the same
reference point.  The expected headers list is common to all
tests.  The source of truth is the `rootfs.Dockerfile` and the
directory contents was generated with:

```bash
docker build -o ./rootfs -f ./rootfs.Dockerfile .
```

The tarball was generated subsequently like so:

```bash
cd rootfs; tar --sort=name -czvf ../rootfs.tar.gz .
```
