+++ initrd/testdata/rootfs/a/b/c/d
@@ -1,0 +1,1 @@
+Hello, World
