+++ initrd/testdata/rootfs/a/b/c/e-symlink
@@ -1,0 +1,1 @@
+./d
