+++ initrd/testdata/rootfs/a/b/c/f-hardlink
@@ -1,0 +1,1 @@
+Hello, World
