+++ initrd/testdata/rootfs/a/b/c/g-recursive-symlink
@@ -1,0 +1,1 @@
+.
