--- initrd/testdata/rootfs/entrypoint.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env sh
-exit 0
