--- initrd/testdata/rootfs/etc/app.conf
@@ -1,2 +1,0 @@
-[app]
-key=value
