--- initrd/testdata/rootfs/lib/libtest.so.1.0.0
@@ -1,1 +1,0 @@
-#fakelib
